Se modifica archivo types y se crea carpeta de tipos
