+++ types/apiResponses.ts
@@ -1,0 +1,4 @@
+export interface EmailWhitelistResponse {
+    is_allowed: boolean;
+    company_id: number | null;
+}
