+++ types/client.ts
@@ -1,0 +1,17 @@
+export type Client = {
+    id: number;
+    id_user: number;
+    height: number;
+    user_image: string;
+    weight: number;
+    weight_unit: string;
+    height_unit: string;
+    streak: number;
+    maximum_streak: number;
+    current_activity: number;
+    target_activity: number;
+    goal: string;
+    created_at: string;
+    updated_at: string;
+    target_weight: number;
+};
