+++ types/coach.ts
@@ -1,0 +1,8 @@
+export type Coach = {
+    id: number;
+    id_user: number;
+    user_image: string;
+    description: string;
+    created_at: string;
+    updated_at: string;
+};
