+++ types/exercise.ts
@@ -1,0 +1,7 @@
+export type Exercise = {
+    id: number;
+    name: string;
+    url_video: string;
+    description: string;
+    created_at: string;
+};
