+++ types/index.ts
@@ -1,0 +1,8 @@
+export * from './user';
+export * from './coach';
+export * from './client';
+export * from './routine';
+export * from './exercise';
+export * from './apiResponses';
+export * from './routineExercise';
+export * from './recipe';
