+++ types/recipe.ts
@@ -1,0 +1,6 @@
+export type Recipe = {
+    id_recipe: number;
+    title: string;
+    content: string;
+    image_url: string;
+}
