+++ types/routine.ts
@@ -1,0 +1,10 @@
+export type Routine = {
+    id: number;
+    id_client: number;
+    id_coach: number;
+    name: string;
+    goal: string;
+    created_at: string;
+    updated: boolean;
+    last_updated: string;
+};
