+++ types/routineExercise.ts
@@ -1,0 +1,11 @@
+export type RoutineExercise = {
+    id: number;
+    id_routine: number;
+    id_exercise: number;
+    sets: number;
+    reps: number;
+    weight: number;
+    note: string;
+    assigned_date: string;
+    created_at: string;
+};
