+++ types/user.ts
@@ -1,0 +1,9 @@
+export type User = {
+    id: number;
+    name: string;
+    id_company: number;
+    id_auth: number;
+    created_at: string;
+    updated_at: string;
+    role: number;
+};
